biohacker/requirements.txt
aws-opentelemetry-distro>=0.10.0
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3
litellm
mcp[cli]
nova-act
opensearch-py
pandas
retrying
strands-agents 
strands-agents-tools[mem0_memory]
streamlit
tqdm
uv
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
bedrock-agentcore
bedrock-agentcore-starter-toolkit
streamlit

opencv-python
pyautogui

# Don't forget to create and activate your environment

# python -m venv .venv

## macOS / Linux: source .venv/bin/activate
## Windows (CMD): .venv\Scripts\activate.bat
## Windows (PowerShell): .venv\Scripts\Activate.ps1
# pip install -r requirements.txt

# Use AWS CLI, a bash script, or just manually set your environment variables
